requirements.txt
openai==1.3.5
tiktoken==0.5.2
anthropic
google-generativeai
python-dotenv==1.0.0
chardet~=5.2.0
boto3~=1.34.84
pytest~=8.1.1